Tweaks
   daogenton.R
   dclftest.R
   deldir.R
   deltametric.R
   [densitybm.R]
   density.lpp.R
   density.ppp.R
   density.psp.R
   derivfv.R
   detpointprocfamilyfun.R
   detPPF-class.R
   dg.R
   dgs.R
   diagnoseppm.R
   diagram.R
   discarea.R
   disc.R
   dist2dpath.R
   distan3D.R
   distances.psp.R
   distances.R
   distanxD.R
   distbdry.R
   distcdf.R
   distfunlpp.R
   distmap.R
   dppm.R
   dummify.R
   dummy.R
	Replaced '1' by '1L' (etc) where appropriate

   DESCRIPTION
   NEWS
	Updated as version 1.48-0.002

+++ R/deldir.R
@@ -42,8 +42,8 @@ delaunay <- function(X) {
   w <- X$window
   dd <- safedeldir(X)
   if(is.null(dd)) return(NULL)
-  a <- dd$delsgs[,5]
-  b <- dd$delsgs[,6]
+  a <- dd$delsgs[,5L]
+  b <- dd$delsgs[,6L]
   use.trigraf  <- get("use.trigraf", envir=.spst.triEnv)
   use.trigrafS <- get("use.trigrafS", envir=.spst.triEnv)
   debug.delaunay <- get("debug.delaunay", envir=.spst.triEnv)
@@ -69,11 +69,11 @@ delaunay <- function(X) {
             ie = as.integer(a),
             je = as.integer(b),
             ntmax = as.integer(ntmax),
-            nt = as.integer(integer(1)),
+            nt = as.integer(integer(1L)),
             it = as.integer(integer(ne)),
             jt = as.integer(integer(ne)),
             kt = as.integer(integer(ne)),
-            status = as.integer(integer(1)))
+            status = as.integer(integer(1L)))
     if(z$status != 0)
       stop("Internal error: overflow in trigrafS")
     tlist <- with(z, cbind(it, jt, kt)[1:nt, ])
@@ -87,11 +87,11 @@ delaunay <- function(X) {
             ie = as.integer(a),
             je = as.integer(b),
             ntmax = as.integer(ntmax),
-            nt = as.integer(integer(1)),
+            nt = as.integer(integer(1L)),
             it = as.integer(integer(ntmax)),
             jt = as.integer(integer(ntmax)),
             kt = as.integer(integer(ntmax)),
-            status = as.integer(integer(1)))
+            status = as.integer(integer(1L)))
     if(z$status != 0)
       stop("Internal error: overflow in trigraf")
     tlist <- with(z, cbind(it, jt, kt)[1:nt, ])
@@ -124,41 +124,41 @@ delaunay <- function(X) {
   # Assemble coordinates of triangles
   x <- X$x
   y <- X$y
-  xtri <- matrix(x[tlist], nrow(tlist), 3)
-  ytri <- matrix(y[tlist], nrow(tlist), 3)
+  xtri <- matrix(x[tlist], nrow(tlist), 3L)
+  ytri <- matrix(y[tlist], nrow(tlist), 3L)
   # ensure triangle vertices are in anticlockwise order
   ztri <- ytri - min(y)
-  dx <- cbind(xtri[,2]-xtri[,1], xtri[,3]-xtri[,2], xtri[,1]-xtri[,3])
-  zm <- cbind(ztri[,1]+ztri[,2], ztri[,2]+ztri[,3], ztri[,3]+ztri[,1])
-  negareas <- apply(dx * zm, 1, sum)
+  dx <- cbind(xtri[,2L]-xtri[,1L], xtri[,3L]-xtri[,2L], xtri[,1L]-xtri[,3L])
+  zm <- cbind(ztri[,1L]+ztri[,2L], ztri[,2L]+ztri[,3L], ztri[,3L]+ztri[,1L])
+  negareas <- apply(dx * zm, 1L, sum)
   clockwise <- (negareas > 0)
   #
   if(any(clockwise)) {
     xc <- xtri[clockwise, , drop=FALSE]
     yc <- ytri[clockwise, , drop=FALSE]
     tc <- tlist[clockwise, , drop=FALSE]
-    xtri[clockwise,]  <- xc[,c(1,3,2)]
-    ytri[clockwise,]  <- yc[,c(1,3,2)]
-    tlist[clockwise,] <- tc[, c(1,3,2)]
+    xtri[clockwise,]  <- xc[,c(1L,3L,2L)]
+    ytri[clockwise,]  <- yc[,c(1L,3L,2L)]
+    tlist[clockwise,] <- tc[, c(1L,3L,2L)]
   }
   # At this point, triangle vertices are listed in anticlockwise order.
   # The same directed edge (i, j) cannot appear twice.
   # To weed out invalid triangles, check for such duplication
-  triedges <- rbind(tlist[, c(1,2)],
-                    tlist[, c(2,3)],
-                    tlist[, c(3,1)])
+  triedges <- rbind(tlist[, c(1L,2L)],
+                    tlist[, c(2L,3L)],
+                    tlist[, c(3L,1L)])
   if(any(bad <- duplicated(triedges))) {
     badedges <- unique(triedges[bad, , drop=FALSE])
     ntri <- nrow(tlist)
     triid <- rep.int(seq_len(ntri), 3)
     illegal <- rep.int(FALSE, ntri)
     for(j in seq_len(nrow(badedges))) {
-      from <- badedges[j, 1]
-      to   <- badedges[j, 2]
+      from <- badedges[j, 1L]
+      to   <- badedges[j, 2L]
       if(debug.delaunay)
         cat(paste("Suspect edge from vertex", from, "to vertex", to, "\n"))
       # find all triangles sharing this edge in this orientation
-      sustri <- triid[(triedges[,1] == from) & (triedges[,2] == to)]
+      sustri <- triid[(triedges[,1L] == from) & (triedges[,2L] == to)]
       if(debug.delaunay)
         cat(paste("\tInvestigating triangles", commasep(sustri), "\n"))
       # list all vertices associated with the suspect triangles
@@ -227,9 +227,9 @@ delaunayDistance <- function(X) {
           adj = as.integer(adj),
           dpath = as.integer(integer(nY * nY)),
           tol = as.integer(0),
-          niter = as.integer(integer(1)), 
-          status = as.integer(integer(1)))
-  if (z$status == -1)
+          niter = as.integer(integer(1L)), 
+          status = as.integer(integer(1L)))
+  if (z$status == -1L)
     warning(paste("graph connectivity algorithm did not converge after", 
                   z$niter, "iterations", "on", nY, "vertices and", 
                   sum(adj) - nY, "edges"))
@@ -310,8 +310,8 @@ delaunayNetwork <- function(X) {
   X <- unique(X, rule="deldir")
   nX <- npoints(X)
   if(nX == 0) return(NULL)
-  if(nX == 1) return(linnet(X, !diag(TRUE)))
-  if(nX == 2) return(linnet(X, !diag(c(TRUE,TRUE))))
+  if(nX == 1L) return(linnet(X, !diag(TRUE)))
+  if(nX == 2L) return(linnet(X, !diag(c(TRUE,TRUE))))
   dd <- safedeldir(X)
   if(is.null(dd)) 
     return(NULL)
